log-service/requirements.txt
Django>=4.2
djangorestframework>=3.14
pytest-django>=4.5
visine-common
